[Feature/#42] 수정된 API연결 및 추가 기능 개발

+++ domain/src/main/java/project/graduation/crowd_sourcing/domain/usecase/MartSearchUseCase.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 class MartSearchUseCase @Inject constructor(
     private val martSearchRepository: MartSearchRepository
 ) {
-    suspend fun searchMartByLocation(lat: Double, lng: Double, radius: Int): List<MartEntity> {
+    suspend fun searchMartByLocation(lat: Double, lng: Double, radius: Double): List<MartEntity> {
         return martSearchRepository.searchMartByLocation(lat, lng, radius)
     }
 
@@ -15,7 +15,7 @@ class MartSearchUseCase @Inject constructor(
         return martSearchRepository.searchMartByZipcode(zipcode, radius)
     }
 
-    suspend fun searchMartByKeyword(keyword: String, radius: Int): List<MartEntity> {
+    suspend fun searchMartByKeyword(keyword: String, radius: Double): List<MartEntity> {
         return martSearchRepository.searchMartByKeyword(keyword, radius)
     }
 }
